backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
alembic>=1.13,<2.0
pydantic-settings>=2.2,<3.0
httpx>=0.27,<1.0
redis>=5.0,<6.0
celery>=5.4,<6.0
python-dotenv>=1.0,<2.0
bcrypt>=5.0,<6.0
python-jose[cryptography]>=3.3,<4.0
